+++ census-angular/src/app/service.service.ts
@@ -45,236 +45,6 @@ export class ServiceService {
     );
   }
 
-  updatePermission(payload: any) {
-    return this.http.patch(
-      `${this.server}/permissions/bulk_update`,
-      payload,
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  updateAccount(account: any): Observable<any> {
-    return this.http.patch(`${this.server}/accounts/${account.id}`, {
-      name: account.name,
-      billing_email: account.billing_email,
-      user_attributes: {
-        id: account.user.id,
-        email: account.user.email,
-        preferred_language: account.user.preferred_language
-      }
-    }, { headers: this.getAuthHeaders() });
-  }
-
-  sendWelcomeEmail(account: any) {
-    return this.http.patch(`${this.server}/accounts/${account.id}/send_welcome_email`,
-      {},
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  toggleGroupActive(groupId: number, isActive: boolean) {
-    return this.http.patch(`${this.server}/groups/${groupId}/toggle_active`,
-      { is_active: isActive },
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  toggleCompanyActive(companyId: number, isActive: boolean) {
-    return this.http.patch(`${this.server}/companies/${companyId}/toggle_active`,
-      { is_active: isActive },
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  getDashboardData() {
-    return this.http.get(
-      `${this.server}/dashboard/show`,
-      {headers: this.getAuthHeaders()}
-    );
-  }
-
-  getMembersByMonth() {
-    return this.http.get(
-      `${this.server}/dashboard/consulta_miembros`,
-      {headers: this.getAuthHeaders()}
-    );
-  }
-
-  getMembersByStateTop5() {
-    return this.http.get(
-      `${this.server}/dashboard/by_state`,
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  updateGroupBenefits(groupId: number, benefits: { id: number; enabled: boolean }[]): Observable<any> {
-    return this.http.patch<any>(
-      `${this.server}/groups/${groupId}/benefits`,
-      { benefits },
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  getTotalEnrolled() {
-    return this.http.get(
-      `${this.server}/dashboard/total_enrolled_last_six_month`,
-      {headers: this.getAuthHeaders()}
-    );
-  }
-
-  getLastetsNotifications() {
-    return this.http.get(
-      `${this.server}/notifications`,
-      {headers: this.getAuthHeaders()}
-    );
-  }
-
-  getAllNotifications() {
-    return this.http.get(
-      `${this.server}/notifications/all`,
-      {headers: this.getAuthHeaders()}
-    );
-  }
-
-  getCounterUnreadNotifications() {
-    return this.http.get(
-      `${this.server}/notifications/count_unread`,
-      {headers: this.getAuthHeaders()}
-    );
-  }
-
-  markAllAsRead() {
-    return this.http.post(
-      `${this.server}/notifications/mark_all_as_read`,
-      {},
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  getAllPageNotifications(page: number = 1) {
-    return this.http.get(
-      `${this.server}/notifications/all?page=${page}`,
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  toggleAdmin(userId: number): Observable<any> {
-    return this.http.patch(
-      `${this.server}/accounts/${userId}/toggle_admin`, 
-      {},
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  markNotificationsAsRead(ids: number[]) {
-    return this.http.post(
-      `${this.server}/notifications/mark_as_read`,
-      { ids },
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  createAccount(accountData: any) {
-    return this.http.post(
-      `${this.server}/accounts/create_with_owner`,
-      accountData,
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  getCompaniesAndGroupsCodifiers() {
-    return this.http.get(
-      `${this.server}/accounts/companies_groups_data`,
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  getAllAccounts(page: number = 1, searchTerm: string = '', sortBy: string = 'name', direction: string = 'asc') {
-    let params: any = {
-      page,
-      q: searchTerm,
-      sort_by: sortBy,
-      direction: direction
-    };
-
-    return this.http.get(`${this.server}/accounts`, {
-      headers: this.getAuthHeaders(),
-      params
-    });
-  }
-
-  deleteAccount(id: number) {
-    return this.http.delete(`${this.server}/accounts/${id}`,
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  getAccountDetails(accountId: number) {
-    return this.http.get(
-      `${this.server}/accounts/${accountId}`,
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  getAccountPermissions(accountId: number, page: number = 1, perPage: number = 10) {
-    return this.http.get(
-      `${this.server}/accounts/${accountId}/permissions?page=${page}&per_page=${perPage}`,
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  getMembersByState() {
-    return this.http.get(`${this.server}/dashboard/members_by_state`,
-      {headers: this.getAuthHeaders()}
-    );
-  }
-
-  removeAdminFromGroup(groupId: number, accountId: number): Observable<void> {
-    return this.http.delete<void>(
-      `${this.server}/groups/${groupId}/admins/${accountId}`,
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  getGroups(page: number = 1, searchTerm: string = '', sort: string = '', direction: string = '') {
-    const params: any = { page };
-
-    if (searchTerm.trim()) params.q = searchTerm.trim();
-    if (sort) params.sort = sort;
-    if (direction) params.direction = direction;
-
-    return this.http.get(`${this.server}/groups`, {
-      headers: this.getAuthHeaders(),
-      params
-    });
-  }
-
-  updateGroup(group: any): Observable<any> {
-    return this.http.put<any>(
-      `${this.server}/groups/${group.id}`,
-      { group: group },
-      { headers: this.getAuthHeaders() }
-    );
-  }
-
-  getGroupDetails(groupId: string): Observable<any> {
-    return this.http.get(`${this.server}/groups/${groupId}`,
-      {headers: this.getAuthHeaders()}
-    );
-  }
-
-  createGroup(groupData: any) {
-    return this.http.post(`${this.server}/groups`, groupData, {
-      headers: this.getAuthHeaders()
-    });
-  }
-
-  getBenefits(): Observable<any[]> {
-    return this.http.get<any[]>(`${this.server}/benefits`, {
-      headers: this.getAuthHeaders()
-    });
-  }
-
